fix: Fix semantic analysis stop after XML query
Fix #706
Change the grammar to process the string literals the same way on the
preprocessor and parser. Add a new rule that accepts incorrect
identifiers on the main lexer to make the mapping find all the required
tokens.
It fixed the semantic analysis by allowing the CobolVisitor to find the
required position after the mapping. Possibly fixes NPE and exceptions
on mapping that could not be reproduced.