refactor: Add possibility to validate statements
All the statements should extend the StatementNode class with their
specific logic. The visit*Statement() method in CobolVisitor should
collect all the localities of the used variables, and then extract the
variables from the map in the validate() method to check their types. If
the variable miss, then it means that it has an invalid defintion and
the according syntax error was created on the previous steps.