refactor: Simplify lookup for nearest locality
Due to the majority of compiler directives is processed on the
preprocessor, there is no need to find the localities for them. The
method for finding the nearest locality was exctracted to a special
class and works only for the embedded languages, if the exact position
for them cannot be found during parsing.