fix: update variables usage which is fully qualified to a single location
Qualification rules
The rules for qualifying a name are:
• A name can be qualified even though it does not need qualification except in a REDEFINES clause, in
which case it must not be qualified.

• Each qualifier must be of a higher level than the name it qualifies and must be within the same
hierarchy.
• If there is more than one combination of qualifiers that ensures uniqueness, any of those combinations
can be used.
• If compiler option QUALIFY(EXTEND) is in effect, and if there is only one fully qualified name that
matches your combination of qualifiers, that reference will be considered unique, even if the set of
qualifiers also matches a partial qualification for a different data item. Fully qualified means every
qualifier is specified.

Signed-off-by: Aman Prashant <aman.prashant@broadcom.com>