Added Syntax Copying for Building Web Extension
Added a line to Build.sh such that the latest syntaxes are copied from the LSP extension to that of the web extension. This is required as webpack needs to be able to include the files, not just the link thereof.

Also added the web extension's syntax folder to the gitignore to avoid duplicating the files.